+++ channel_integrations/integrated_channel/constants.py
@@ -4,3 +4,5 @@ Constants used by the integrated channels
 
 ISO_8601_DATE_FORMAT = '%Y-%m-%dT%H:%M:%S.%fZ'
 TASK_LOCK_EXPIRY_SECONDS = 60 * 60 * 12
+
+TPA_ORG_ALLOWLIST_ADMIN_ROLE = 'tpa_org_allowlist_admin'